rename asg to mmi_bigram (#70)
