+++ egs/librispeech/asr/simple_v1/run.sh
@@ -44,10 +44,10 @@ fi
 
 if [ $stage -le 6 ]; then
   # python3 ./train.py # ctc training
-  python3 ./asg_train.py
+  python3 ./mmi_bigram_train.py
 fi
 
 if [ $stage -le 7 ]; then
   # python3 ./decode.py # ctc decoding
-  python3 ./asg_decode.py
+  python3 ./mmi_bigram_decode.py
 fi
